refactor: move ProgramHandler inside of PreflightInterpretedInstance
